+++ controller/MateriaController.php
@@ -3,11 +3,13 @@ require_once "model\MateriaModel.php";
 require_once "view\MateriaView.php";
 require_once "helpers/AuthHelper.php";
 require_once "model/CarreraModel.php";
+require_once "view\UserView.php";
 //SUBJECT = MATERIA  
 class MateriaController
 {
     private $model;
     private $view;
+    private $view_user;
     private $helper;
     private $carrera_model;
 
@@ -16,6 +18,7 @@ class MateriaController
         $this->model = new MateriaModel();
         $this->carrera_model = new CarreraModel();
         $this->view = new MateriaView();
+        $this->view_user = new UserView();
         $this->helper = new AuthHelper();
     }
     //filtrar materias
@@ -37,8 +40,9 @@ class MateriaController
     public function formSubject()
     {
 
+        $isAdmin = $this->helper->checkLoggedIn();
         $carreras = $this->carrera_model->getDegreeProgram();
-        $this->view->renderFormSubject($carreras);
+        $this->view->renderFormSubject($carreras, $isAdmin);
     }
 
 
@@ -73,26 +77,32 @@ class MateriaController
     public function showTableOfSubjects()
     {
 
-
+        $isAdmin = $this->helper->checkLoggedIn();
         $tablasMaterias = $this->model->getTableOfSubjects();
-        $this->view->renderTableSubjects($tablasMaterias);
+        $this->view->renderTableSubjects($tablasMaterias, $isAdmin);
     }
     //   BORRAR MATERIA
     public function deleteSubject($id)
     {
-        $this->model->deleteSubject($id);
-        $this->view->renderTableOfLocationSubjects();
-
-        $this->view->renderTableOfLocationSubjects();
+        $isAdmin = $this->helper->checkLoggedIn();
+        if ($isAdmin == true) {
+            $this->model->deleteSubject($id);
+            $this->view->renderTableOfLocationSubjects();
+        } else {
+            $this->view_user->renderLogin();
+        }
     }
 
     //EDITAR MATERIA
     public function editSubject($id_materia)
     {
-
-        $this->model->editSubject($_POST['nombre'], $_POST['profesor'], $_POST['id_carrera'], $id_materia);
-        $this->view->renderTableOfLocationSubjects();
-        
+        $isAdmin = $this->helper->checkLoggedIn();
+        if ($isAdmin == true) {
+            $this->model->editSubject($_POST['nombre'], $_POST['profesor'], $_POST['id_carrera'], $id_materia);
+            $this->view->renderTableOfLocationSubjects();
+        } else {
+            $this->view_user->renderLogin();
+        }
     }
     public function redirectHome()
     {
